# backend-flask/api/open_api/openapi.yaml
openapi: 3.0.3
info:
  title: "SC API pos"
  description: "Модуль для интеграции с кассами для лояльности на смарт контрактах."
  version: 1.0.0
servers:
  - url: https://dm.api/20130701/
    description: "Production сервер"
  - url: http://staging-api.example.com
    description: "Staging сервер"
...
paths:
  /tokens/:
    post:
      summary: "Получение нового токена"
      operationId: src.tokens.views.create_token
      tags:
      - Tokens
      description: "Запрос на получение нового токена."
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                merchant_shop:
                  $ref: '#/components/schemas/merchant_shop'
                pos:
                  $ref: '#/components/schemas/pos'
                description:
                  $ref: '#/components/schemas/description'

      responses:
        '201':
          description: "Метод возвращает значение токена авторизации, которое затем следует использовать в качестве значения параметра id метода GET этого ресурса и для авторизации при вызове методов других ресурсов. \n\nTокен авторизации приложения должен быть указан в каждом запросе к API."
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    $ref: '#/components/schemas/token'
                  token_url:
                    $ref: '#/components/schemas/token_url'
        '401':
          description: "При обращении к ресурсу без указания токена или с указанием неверного токена возвращается пустой ответ с кодом ошибки 401 Unauthorized."
        '409':
          description: "Если для указанного в запросе участника системы (параметр merchant_shop) и кассового терминала (параметр pos) токен авторизации уже был выдан, но ещё не активирован, то будет возвращён статус HTTP ответа 409 Conflict и текстовое сообщение, описывающее ошибочную ситуацию (например, «Для этого терминала уже есть не активированный токен.»)."
          content:
            application/json:
              schema:
                type: object
                properties:
                  TODO:
                    type: string
                    description: "Нет примера ответа в документации"
                    example: "Нет примера ответа в документации"
  /tokens/{token}:
    get:
      summary: "Статус токена авторизации pos"
      operationId: src.tokens.views.get_token
      tags:
      - Tokens
      description: "Возвращает статус активации токена. В случае если токен авторизации активирован будет возвращён статус HTTP 200 OK, в ином случае 404 Not Found."
      parameters:
      - name: token
        in: path
        required: true
        description: "Токен авторизации pos"
        schema:
          $ref: '#/components/schemas/token'
      responses:
        '200':
          description: "Если токен авторизации активирован будет получен соответствующий ответ."
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                    example: true
        '404':
          description: "В другом любом случае."
          content:
            application/json:
              schema:
                type: object
                properties:
                  TODO:
                    type: string
                    example: "Нет примера ответа в документации"

    delete:
      summary: "Отзыв (отключение) токена авторизации pos"
      operationId: src.tokens.views.delete_token
      tags:
      - Tokens
      description: "Отзыв (отключение) полученного ранее токена. После успешного выполнения возвращается статус HTTP 204 No Content, в ином случае 404 Not Found."
      parameters:
      - name: token
        in: path
        required: true
        description: "Токен авторизации pos"
        schema:
          $ref: '#/components/schemas/token'
      responses:
        '204':
          description: "Успешное выполнение."
        '404':
          description: "В другом любом случае."
...
components:
  securitySchemes:
    bearerAuthEvotor:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_evotor

    bearerAuthAdmin:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_adminsc

    bearerAuthPos:
      type: apiKey
      in: header
      name: Dmpos-Authorization
      x-apikeyInfoFunc: src.middleware.bearer_auth_pos
...
  schemas:
...
    merchant_shop:
      type: string
      description: "ID (строка цифр длиной до 25)\n\nИдентификатор участника системы - владельца терминала."
      maxLength: 25
    pos:
      type: string
      description: "Строка длиной до 25.\n\nНомер терминала (кассы)."
      maxLength: 25
      example: "1"
    description:
      type: string
      description: "Строка длиной до 100.\n\nОписание терминала (кассы) - не обязательно."
      maxLength: 100
    token:
      type: string
      description: "Токен авторизации, которой следует использовать в качестве значения параметра id при вызове методов."
      example: "dcb0cb4e77164e60612c162745b7072fbefeaba3"
    token_url:
      type: string
      description: "Ссылка на ресурс созданного токена авторизации."
      example: "http://dm.api/20130701/tokens/dcb0cb4e77164e60612c162745b7072fbefeaba3"
...